* watches: Make it possible to enter new watches with the keyboard only in wx>=29
The wx28 version of the code depends on a strange feature of wxpropgrid,
that sends property selected message after the call SelectProperty. This is
no longer the case with wxpropgrid integrated in wx29. To fix the issue we
have to call BeginLabelEdit right in the call that adds the empty property.

git-svn-id: http://svn.code.sf.net/p/codeblocks/code/trunk@9571 2a5c6006-c6dd-42ca-98ab-0921f2732cef