* MacrosManager - Add additional tests to MacrosManager to determine if a newer project has been activated.
When a workspace with only one project is closed, then another loaded,
or another project in a multi-project workspace is compiled, 
it's possible for the newly loaded/referenced project to be assigned the same memory address as the previous project.
Any test against the previously loaded projects address will match, and
macrosmanager.cpp will not replace correctly the macros for the second project.

git-svn-id: http://svn.code.sf.net/p/codeblocks/code/trunk@10898 2a5c6006-c6dd-42ca-98ab-0921f2732cef